+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+1.1.0
+
+- Use special endpoint for waterservices data > 120 days
+- Update dates of last data update
+- Adjust "fair" rating for turbidity
+- Update variable descriptions with client-provided copy
+
 1.0.0
 
 - Initial production release